# UaRestGateway/uarestgateway.client/public/opcfoundation-org-opcua-rest-api-1.0.0-swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OPC UA REST API",
    "description": "This API provides simple REST based access to an OPC UA server.",
    "contact": {
      "email": "office@opcfoundation.org"
    },
    "license": {
      "name": "OPC Source Deliverable Agreement of Use",
      "url": "https://opcfoundation.org/license/source/1.11/"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://localhost:7010"
    }
  ],
  "paths": {
    "/opcua/read": {
      "post": {
        "operationId": "Read",
        "requestBody": {
          "description": "ReadRequestMessage",
          "content": {
            "application/json": {
              "schema": {
                "title": "ReadRequestMessage",
                "properties": {
                  "NamespaceUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServerUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "LocaleIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServiceId": {
                    "type": "integer",
                    "format": "uint32"
                  },
                  "Body": {
                    "$ref": "#/components/schemas/ReadRequest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ReadResponseMessage",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ReadResponseMessage",
                  "properties": {
                    "NamespaceUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServerUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServiceId": {
                      "type": "integer",
                      "format": "uint32"
                    },
                    "Body": {
                      "$ref": "#/components/schemas/ReadResponse"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/opcua/write": {
      "post": {
        "operationId": "Write",
        "requestBody": {
          "description": "WriteRequestMessage",
          "content": {
            "application/json": {
              "schema": {
                "title": "WriteRequestMessage",
                "properties": {
                  "NamespaceUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServerUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "LocaleIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServiceId": {
                    "type": "integer",
                    "format": "uint32"
                  },
                  "Body": {
                    "$ref": "#/components/schemas/WriteRequest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "WriteResponseMessage",
            "content": {
              "application/json": {
                "schema": {
                  "title": "WriteResponseMessage",
                  "properties": {
                    "NamespaceUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServerUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServiceId": {
                      "type": "integer",
                      "format": "uint32"
                    },
                    "Body": {
                      "$ref": "#/components/schemas/WriteResponse"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/opcua/call": {
      "post": {
        "operationId": "Call",
        "requestBody": {
          "description": "CallRequestMessage",
          "content": {
            "application/json": {
              "schema": {
                "title": "CallRequestMessage",
                "properties": {
                  "NamespaceUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServerUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "LocaleIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServiceId": {
                    "type": "integer",
                    "format": "uint32"
                  },
                  "Body": {
                    "$ref": "#/components/schemas/CallRequest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "CallResponseMessage",
            "content": {
              "application/json": {
                "schema": {
                  "title": "CallResponseMessage",
                  "properties": {
                    "NamespaceUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServerUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServiceId": {
                      "type": "integer",
                      "format": "uint32"
                    },
                    "Body": {
                      "$ref": "#/components/schemas/CallResponse"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/opcua/historyread": {
      "post": {
        "operationId": "HistoryRead",
        "requestBody": {
          "description": "HistoryReadRequestMessage",
          "content": {
            "application/json": {
              "schema": {
                "title": "HistoryReadRequestMessage",
                "properties": {
                  "NamespaceUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServerUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "LocaleIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServiceId": {
                    "type": "integer",
                    "format": "uint32"
                  },
                  "Body": {
                    "$ref": "#/components/schemas/HistoryReadRequest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "HistoryReadResponseMessage",
            "content": {
              "application/json": {
                "schema": {
                  "title": "HistoryReadResponseMessage",
                  "properties": {
                    "NamespaceUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServerUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServiceId": {
                      "type": "integer",
                      "format": "uint32"
                    },
                    "Body": {
                      "$ref": "#/components/schemas/HistoryReadResponse"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/opcua/historyupdate": {
      "post": {
        "operationId": "HistoryUpdate",
        "requestBody": {
          "description": "HistoryUpdateRequestMessage",
          "content": {
            "application/json": {
              "schema": {
                "title": "HistoryUpdateRequestMessage",
                "properties": {
                  "NamespaceUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServerUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "LocaleIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServiceId": {
                    "type": "integer",
                    "format": "uint32"
                  },
                  "Body": {
                    "$ref": "#/components/schemas/HistoryUpdateRequest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "HistoryUpdateResponseMessage",
            "content": {
              "application/json": {
                "schema": {
                  "title": "HistoryUpdateResponseMessage",
                  "properties": {
                    "NamespaceUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServerUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServiceId": {
                      "type": "integer",
                      "format": "uint32"
                    },
                    "Body": {
                      "$ref": "#/components/schemas/HistoryUpdateResponse"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/opcua/browse": {
      "post": {
        "operationId": "Browse",
        "requestBody": {
          "description": "BrowseRequestMessage",
          "content": {
            "application/json": {
              "schema": {
                "title": "BrowseRequestMessage",
                "properties": {
                  "NamespaceUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServerUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "LocaleIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServiceId": {
                    "type": "integer",
                    "format": "uint32"
                  },
                  "Body": {
                    "$ref": "#/components/schemas/BrowseRequest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "BrowseResponseMessage",
            "content": {
              "application/json": {
                "schema": {
                  "title": "BrowseResponseMessage",
                  "properties": {
                    "NamespaceUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServerUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServiceId": {
                      "type": "integer",
                      "format": "uint32"
                    },
                    "Body": {
                      "$ref": "#/components/schemas/BrowseResponse"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/opcua/browsenext": {
      "post": {
        "operationId": "BrowseNext",
        "requestBody": {
          "description": "BrowseNextRequestMessage",
          "content": {
            "application/json": {
              "schema": {
                "title": "BrowseNextRequestMessage",
                "properties": {
                  "NamespaceUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServerUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "LocaleIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServiceId": {
                    "type": "integer",
                    "format": "uint32"
                  },
                  "Body": {
                    "$ref": "#/components/schemas/BrowseNextRequest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "BrowseNextResponseMessage",
            "content": {
              "application/json": {
                "schema": {
                  "title": "BrowseNextResponseMessage",
                  "properties": {
                    "NamespaceUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServerUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServiceId": {
                      "type": "integer",
                      "format": "uint32"
                    },
                    "Body": {
                      "$ref": "#/components/schemas/BrowseNextResponse"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/opcua/translate": {
      "post": {
        "operationId": "TranslateBrowsePathsToNodeIds",
        "requestBody": {
          "description": "TranslateBrowsePathsToNodeIdsRequestMessage",
          "content": {
            "application/json": {
              "schema": {
                "title": "TranslateBrowsePathsToNodeIdsRequestMessage",
                "properties": {
                  "NamespaceUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServerUris": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "LocaleIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ServiceId": {
                    "type": "integer",
                    "format": "uint32"
                  },
                  "Body": {
                    "$ref": "#/components/schemas/TranslateBrowsePathsToNodeIdsRequest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "TranslateBrowsePathsToNodeIdsResponseMessage",
            "content": {
              "application/json": {
                "schema": {
                  "title": "TranslateBrowsePathsToNodeIdsResponseMessage",
                  "properties": {
                    "NamespaceUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServerUris": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ServiceId": {
                      "type": "integer",
                      "format": "uint32"
                    },
                    "Body": {
                      "$ref": "#/components/schemas/TranslateBrowsePathsToNodeIdsResponse"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BrowseRequest": {
        "type": "object",
        "properties": {
          "RequestHeader": {
            "$ref": "#/components/schemas/RequestHeader"
          },
          "View": {
            "$ref": "#/components/schemas/ViewDescription"
          },
          "RequestedMaxReferencesPerNode": {
            "type": "integer",
            "format": "uint32"
          },
          "NodesToBrowse": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrowseDescription"
            }
          }
        }
      },
      "BrowseResponse": {
        "type": "object",
        "properties": {
          "ResponseHeader": {
            "$ref": "#/components/schemas/ResponseHeader"
          },
          "Results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrowseResult"
            }
          },
          "DiagnosticInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticInfo"
            }
          }
        }
      },
...
      "TranslateBrowsePathsToNodeIdsRequest": {
        "type": "object",
        "properties": {
          "RequestHeader": {
            "$ref": "#/components/schemas/RequestHeader"
          },
          "BrowsePaths": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrowsePath"
            }
          }
        }
      },
...
      "ReadRequest": {
        "type": "object",
        "properties": {
          "RequestHeader": {
            "$ref": "#/components/schemas/RequestHeader"
          },
          "MaxAge": {
            "type": "number",
            "format": "double"
          },
          "TimestampsToReturn": {
            "type": "integer",
            "format": "TimestampsToReturn"
          },
          "NodesToRead": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReadValueId"
            }
          }
        }
      },
      "ReadResponse": {
        "type": "object",
        "properties": {
          "ResponseHeader": {
            "$ref": "#/components/schemas/ResponseHeader"
          },
          "Results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataValue"
            }
          },
          "DiagnosticInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticInfo"
            }
          }
        }
      },
...
      "HistoryReadRequest": {
        "type": "object",
        "properties": {
          "RequestHeader": {
            "$ref": "#/components/schemas/RequestHeader"
          },
          "HistoryReadDetails": {
            "$ref": "#/components/schemas/ExtensionObject"
          },
          "TimestampsToReturn": {
            "type": "integer",
            "format": "TimestampsToReturn"
          },
          "ReleaseContinuationPoints": {
            "type": "boolean"
          },
          "NodesToRead": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoryReadValueId"
            }
          }
        }
      },
      "HistoryReadResponse": {
        "type": "object",
        "properties": {
          "ResponseHeader": {
            "$ref": "#/components/schemas/ResponseHeader"
          },
          "Results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoryReadResult"
            }
          },
          "DiagnosticInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticInfo"
            }
          }
        }
      },
...
      "WriteRequest": {
        "type": "object",
        "properties": {
          "RequestHeader": {
            "$ref": "#/components/schemas/RequestHeader"
          },
          "NodesToWrite": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WriteValue"
            }
          }
        }
      },
      "WriteResponse": {
        "type": "object",
        "properties": {
          "ResponseHeader": {
            "$ref": "#/components/schemas/ResponseHeader"
          },
          "Results": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "StatusCode"
            }
          },
          "DiagnosticInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticInfo"
            }
          }
        }
      },
...
      "HistoryUpdateRequest": {
        "type": "object",
        "properties": {
          "RequestHeader": {
            "$ref": "#/components/schemas/RequestHeader"
          },
          "HistoryUpdateDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExtensionObject"
            }
          }
        }
      },
      "HistoryUpdateResponse": {
        "type": "object",
        "properties": {
          "ResponseHeader": {
            "$ref": "#/components/schemas/ResponseHeader"
          },
          "Results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoryUpdateResult"
            }
          },
          "DiagnosticInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticInfo"
            }
          }
        }
      },
...
      "CallRequest": {
        "type": "object",
        "properties": {
          "RequestHeader": {
            "$ref": "#/components/schemas/RequestHeader"
          },
          "MethodsToCall": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CallMethodRequest"
            }
          }
        }
      },
      "CallResponse": {
        "type": "object",
        "properties": {
          "ResponseHeader": {
            "$ref": "#/components/schemas/ResponseHeader"
          },
          "Results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CallMethodResult"
            }
          },
          "DiagnosticInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticInfo"
            }
          }
        }
      },
...
    }
  }
}